requirements.txt
# Depo-Pro Transcribe — dependencies
# Install: pip install -r requirements.txt

# UI
customtkinter>=5.2.0

# Deepgram transcription
deepgram-sdk>=6,<7

# HTTP client (required by deepgram-sdk)
httpx>=0.27.0

# Document handling
python-docx>=1.1.0
pdfplumber>=0.10.0
reportlab>=4.0.0

# Environment / config
python-dotenv>=1.0.1

# Testing
pytest>=8.0.0
